Update 42-soft-saturation.rst
I found this repo while looking into the daisy DSP library, which reportedly has some glitches. While looking into it, I found that they implemented this algorithm but they missed the fact that the case x=a is not handled. I'm fixing it there, but I thought having the change here would also be of help to other people looking into this algorithm